Simplified
I think I over engineered it, so I decided it would be a good idea to simplify it. Now you just write true or false to get_output function, and in case if you wrote "true" the result will be ranging from -1 to 1 (this will not work if you are using function that gives output not ranging 0 to 1), in case you wrote "false" nothing will change and outout will be from 0 to 1.